Changes carbonicity to a double (like it's supposed to be

+++ base9/Settings.hpp
@@ -17,7 +17,7 @@ extern "C"
     struct WhiteDwarfSettings
     {
         int filterSet;
-        int carbonicity;
+        double carbonicity;
     };
 
     struct BrownDwarfSettings
